update on problem
small update to capture though on solving the hault issue on swipe out.
increase total number of leds the program thinks are available by 38.
i believe that the program is stopping swipe out at led 72 as it is the max total of leds available to calculate the animation.
by increasing the total number by the number of leds used for overflow i believe this will solve the issue.
eg. swipe in works because it is not copying the right to the left, therefore +36 leds work to hide the animation, as that is the number of leds available but are being copied over from the left. hence creating the fade out effect in the center. i however am not sure how to use these new variables in the program